tests/integration/test_workflows.py
"""
Integration tests for Agent3D core workflows.
"""
import pytest
from pathlib import Path
import yaml
import os

class TestCoreWorkflows:
    """Test core Agent3D workflows."""

    @pytest.fixture
    def sample_config(self, tmp_path):
        """Create a sample configuration file."""
        config = {
            'project': {
                'name': 'test_project',
                'type': 'test',
                'language': 'python'
            },
            'identifier_patterns': {
                'TC-': {
                    'pattern': 'TC-.*',
                    'primary_files': ['docs/features/*.md'],
                    'reference_files': ['**/*.py']
                }
            }
        }
        config_path = tmp_path / '.agent3d-config.yml'
        with open(config_path, 'w') as f:
            yaml.dump(config, f)
        return config_path

    def test_config_loading(self, sample_config):
        """Test that configuration files are loaded correctly."""
        assert sample_config.exists(), "Config file was not created"
        with open(sample_config) as f:
            config = yaml.safe_load(f)
        assert config['project']['name'] == 'test_project'
        assert 'TC-' in config['identifier_patterns']

    def test_workflow_initialization(self, tmp_path):
        """Test basic workflow initialization."""
        # This is a placeholder for actual workflow initialization test
        # Will be expanded once core workflow implementation is available
        assert True

    @pytest.mark.slow
    def test_full_workflow_execution(self, tmp_path):
        """Test execution of a full workflow."""
        # This is a placeholder for end-to-end workflow test
        # Will be expanded once workflow implementation is complete
        assert True
