requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
coverage>=6.0.0
pyyaml>=6.0.0
pytest-html>=3.0.0
